box: expose memtx_engine_recovery_state
Recovery state of memtx engine is a key state of entire tarantool.
Knowledge of that state allows to distinguish whether a space or
a secondary index is absent because of user error or because the
recovery did not come to its creation.

It also should be noted that only one instance of memtx engine is
allowed and thus there could be no collision if we export memtx
recovery state as a separate global variable.

Extract the state from memtx engine structure to global variable.

NO_DOC=refactoring
NO_CHANGELOG=refactoring