salad: normalize distance calculation in rtree
Originally SOP_NEIGHBOR iterator in rtree was made in a strange
manner: it takes given rectangle but uses only top-left point of
it, ignoring the rest coordinates. Internally this means that a
special function required that calculated distance from any rect
to `neighbor` rect, from which the top-left point is taken.

On the other hand in tarantool the point from which the KNN search
takes place is always set as a correct point, that is a rectangle
with zero width and height.

So there will be no harm if we replace special `neighbor` function
with general function that calculate distance between arbitrary
rectangles. That makes the code more simple and easy to support.

This patch implements this.

NO_DOC=refactoring
NO_CHANGELOG=refactoring