txm: carefully handle conflict
When a trasnaction is in read-confirmed state it must ignore all
prepared changes, and if it actually ignores something - it must
fall to read-view state.

By a mistake the check relied not on actual skipping of a prepared
statement, but on the fact that there is a deleting statement.

Fix it by raising a conflict only if a deleting statement is skipped.

NO_DOC=bugfix