memtx: refactor mvcc story linking to the top of chain
Before this patch there were several different places in the code
that deal with referencing tuple in space, setting in_index member
and marking the story as retained or not. But logically all above
is about the same - about placing a story to the top of a chain,
i.e. the first story in version list to which index points.

This commit refactors these things a bit. This mostly relates to
two functions - memtx_tx_story_new and memtx_tx_story_link_top.

Changes in memtx_tx_story_new are based on the fact that if a story
is created by tuple, it is or immediately will be at the top of
chain. Considering this we can omit argument `is_referenced_to_pk`
and always create a story ready to be in top of chain. If a story
is already in the top - nothing esle is need; if it is to become
the top - memtx_tx_story_link_top must be called after.

Further, linking to top of chain is needed exactly in two cases:
* if a story just created by memtx_tx_story_new must become a top
* if a chain is reordered involving the top story (the top and the
  next stories are swapped)
These two cases are logically very close but still different.
Even more, previously there ware two functions for that:
memtx_tx_story_new_light and memtx_tx_story_new correspondingly.
This commit introduces one function for that (although with one
more argument) that also incapsulates activities about referencing
tuples and marking stories as retained.

After this patch the rules are logical and simple:
* if a tuple is inserted - call _story_new and _link_top.
* if a story of existring tuple is needed - call _story_new.
* if chain is reordered involving top story - call _link_top.